pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autolunch"
version = "1.0.0"
description = "Autonomous Zomato lunch ordering agent with LLM decision engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core LLM / AI
    "openai>=1.30.0",           # OpenRouter uses OpenAI-compatible SDK

    # Data validation & settings
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # HTTP
    "httpx>=0.27.0",

    # Logging
    "loguru>=0.7.2",

    # Google integrations
    "gspread>=6.1.0",
    "google-auth>=2.30.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "rich>=13.7.0",             # Pretty CLI output for dry-run testing
    "tenacity>=8.3.0",          # Retry logic for API calls
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
]

[project.scripts]
autolunch = "autolunch.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
